# Optimisation (?) in cursor: Use a generator instead of a list