@Starbuck5 2021-03-06
Based on your tests, it looks like using a list instead of a generator would be more performant for all practical applications of making a bitmap type cursor.

That being said, I don't have a preference which way to go. Cursor objects don't have much need to be performant.

I wrote this code, and I only did a list comprehension because I didn't know the other way would work :).